Handle global variables and external functions
The former is necessary for the latter because our tests pass global (constant)
strings to printf. The expressions for globals are stored in "shadow" globals.
At function call sites, we now always push the expressions for the arguments and
assume that the function sets an expression for the return value. This will
break for external functions that haven't been compiled with our pass; we need
some mechanism to verify whether the function has actually set an expression.