Switch to renaming instead of redefinition for interception
Redefining functions with the same name has the problem that we need to be
extremely careful about visibility: only the code that we compiled should ever
see the interceptors. However, some functions' visibility can't be changed,
e.g., malloc.

The new strategy consists of prefixing every function name with a known string;
for libc functions, we then define wrappers around the original functions with
the expected name.