Fix a bug in the handling of "switch" instructions
The target block of a switch case can have other incoming edges; our handling
would change the predecessor from the block containing the switch to some newly
generated blocks, which confuses phi nodes if there are any. The fix is to push
path constraints according to the cases of the switch instruction before the
actual switch. While it adds a number of checks and pushes equal to the number
of constraints, it simplifies the handling a lot and keeps code size manageable.