Don't call the backend for every intrinsic
LLVM intrinsic functions don't really count as functions for the purposes of
context-aware pruning, and in many cases we don't need to handle them. This
commit avoids generating unnecessary backend calls in such situations; it should
be particularly helpful for the performance of code that is compiled with debug
information (which generates lots of calls to intrinsics).