Make tests endianness-agnostic
Currently tests take binary input in little-endian format and thus fail
on big-endian systems.

Fix by converting inputs to network byte order, using ntohl() in tests
and providing the symbolic ntohl() wrapper. Other functions from this
family could be similarly added as well, but it is not required for
this fix.