test: fix tests
This fixes issue #29 and ensures that tests pass.

The issue was that the GNU netinet/in.h header files replaces ntohl with
a byteswap macro when optimizations are enabled. Thus, we check if ntohl
is defined as a macro, and if it is, undefine the macro to ensure that
the function is used (and hence symbolized).