Merge upstream (#3)
* This is a temporary fix due to std::iterator depercation.

This commit needs to be reverted once a proper fix is in place.

* symcc_fuzzing_helper: Move to clap3 (#94)

* Revert "symcc_fuzzing_helper: Move to clap3 (#94)" (#101)

This reverts commit 88b464c290786b58e5883f5aa1d08e2a3d5de0fb.

* Add some FAQs to the Readme

* changed from structopt to clap 3 (#103)

* fix for issue #108

* fix for issue #108

* LLVM 12 works without changes

* Add a clang-format configuration

This is just the output of "clang-format -style=llvm -dump-config".

* Add support for LLVM 13

Clang now uses the new pass manager for the optimization pipeline, so we
have to do the same to make Clang use our pass. Moreover, FileCheck now
complains if a configured prefix doesn't appear in the checked file; added
"ANY" in three tests where it was missing. Finally, printing
arbitrary-precision integers in QSYM needed some changes.

* Add support for LLVM 14

* LLVM 15 works without changes

* fix issue #109

* Run clang-format

We should really automate this...

* Add a GitHub action that checks LLVM compatibility

* Prevent test failures in case of reordered solver output

Z3 doesn't always output model constants in the same order; make sure
that our tests don't depend on it.

* Accept symbolic input from memory

This commit adds the option to mark symbolic input by calling
symcc_make_symbolic from the program under test.

The refactoring that was required to add the new feature has had the
pleasant side effect that the QSYM backend now doesn't require the
entire input upfront anymore, making it much more convenient to feed
symbolic data through stdin.

* Run GitHub actions for pull requests only

No need for "push": the "pull_request" event already triggers when new
commits are pushed to the PR branch, and we expect all changes to go
through a PR.

Co-authored-by: Aurelien Francillon <aurelien.francillon@eurecom.fr>
Co-authored-by: Dominik Maier <domenukk@gmail.com>
Co-authored-by: aurelf <aurelien@francillon.net>
Co-authored-by: Dominik Maier <dmnk@google.com>
Co-authored-by: Emilio Coppa <ercoppa@gmail.com>
Co-authored-by: Sebastian Poeplau <poeplau@adacore.com>