Fix handling of expressions for floats and Booleans in structs
This commit fixes two problems, both part of eurecom-s3/symcc#138:

1. When creating expressions for struct literals, we now convert
expressions for floats and Booleans to bit-vector kind before attempting
to concatenate them with the rest of the struct expression.

2. In "insertvalue" instructions with a symbolic target and a concrete
value to insert we now make sure that the runtime call to create the
expression for the inserted value isn't optimized out.