Fix handling of the llvm.eh.typeid.for intrinsic
We use IntrinsicLowering to turn a number of supported intrinsics into
code or library calls that we can instrument. This transformation is
expected to preserve code semantics. However, IntrinsicLowering handles
llvm.eh.typeid.for by inserting a dummy value of constant 1, which
changes code semantics in many cases (except when 1 happens to be the
expected type ID).

This commit makes the compiler stop using IntrinsicLowering for
llvm.eh.typeid.for; instead, we treat the intrinsic's result as a
constant.

Fixes eng/das/fuzz/symcc#23.